requirements.txt
# Core web framework
fastapi==0.115.1
uvicorn==0.34.2

# LangChain integrations
langchain==0.3.24
langchain-openai==0.3.14
langchain-pinecone==0.2.5
langchain-core==0.3.56

# Pinecone and OpenAI clients
pinecone==6.0.2
openai==1.76.0

# Token counting
tiktoken==0.9.0

# .env support
python-dotenv==1.1.0

# Pydantic (used explicitly)
pydantic==2.11.3
pydantic-settings==2.9.1